refactor: EditImageSectionをコンポーネント化
